refactor: refactor test package, and add empty test files
