+++ src/test/java/bowling/FrameScoreTests.java
@@ -1,62 +1,54 @@
 package bowling;
 
+import bowling.FrameScore;
+import bowling.LeftScoreCount;
+import bowling.Score;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import java.util.Arrays;
-
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 
-@DisplayName("FrameScoreTests")
+@DisplayName("프레임 스코어 테스트")
 public class FrameScoreTests {
 
-    @DisplayName("FrameScore 생성 테스트")
+    @DisplayName("생성 테스트")
     @Test
-    public void generateFrameScoreTest() {
-        assertThatCode(FrameScore::new);
+    public void generateTest() {
+        assertThatCode(() -> FrameScore.newInstance(Score.of(7), LeftScoreCount.of(1)));
     }
 
-    @DisplayName("FrameScore add 테스트")
+    @DisplayName("생성 (Miss) 테스트")
     @Test
-    public void addFrameScoreTest() {
-        FrameScore frameScore = new FrameScore();
-        assertThatCode(() -> frameScore.add(5));
-        assertThatCode(() -> frameScore.add(4));
+    public void generateMissTest() {
+        assertThatCode(() -> FrameScore.createMiss(Score.of(7)));
     }
 
-    @DisplayName("FrameScore sum 테스트")
+    @DisplayName("생성 (Spare) 테스트")
     @Test
-    public void sumFrameScoreTest() {
-        FrameScore frameScore = FrameScore.newInstance(Arrays.asList(1, 2));
-        assertThat(frameScore.sum()).isEqualTo(3);
+    public void generateSpareTest() {
+        assertThatCode(FrameScore::createSpare);
     }
 
-    @DisplayName("FrameScore count 테스트")
+    @DisplayName("생성 (Strike) 테스트")
     @Test
-    public void countFrameScoreTest() {
-        FrameScore frameScore = FrameScore.newInstance(Arrays.asList(1, 2));
-        assertTrue(frameScore.isSameScoreCount(2));
+    public void generateStrikeTest() {
+        assertThatCode(FrameScore::createStrike);
     }
 
-    @DisplayName("SubTotalScore 반환 테스트")
+    @DisplayName("스코어 반환 테스트")
     @Test
-    public void getSubTotalScoreTest() {
-        FrameScore frameScore = FrameScore.newInstance(Arrays.asList(10));
-        SubTotal expectedSubTotal = SubTotal.newInstance(10, NextAddingUpScores.newInstance(Arrays.asList(Score.of(10))));
-
-        assertThat(frameScore.getSubTotal()).isEqualTo(expectedSubTotal);
+    public void getScoreTest() {
+        FrameScore frameScore = FrameScore.createSpare();
+        assertThat(frameScore.getScore()).isEqualTo(Score.of(10));
     }
 
-    @DisplayName("SubTotalScore 반환 테스트2")
+    @DisplayName("스코어 계산 가능 체크 테스트")
     @Test
-    public void getSubTotalScoreTest2() {
-        FrameScore frameScore = FrameScore.newInstance(Arrays.asList(10));
-        NextAddingUpScores nextAddingUpScores = NextAddingUpScores.newInstance(Arrays.asList(Score.of(10), Score.of(10)));
-        SubTotal expectedSubTotal = SubTotal.newInstance(30, NextAddingUpScores.newInstance(Arrays.asList(Score.of(10), Score.of(10))));
-
-        assertThat(frameScore.getSubTotal(nextAddingUpScores)).isEqualTo(expectedSubTotal);
+    public void canCalculateScoreTest() {
+        FrameScore frameScore = FrameScore.createMiss(Score.of(7));
+        assertFalse(frameScore.canCalculateSelfScore());
     }
 
 }
