+++ src/test/java/bowling/frame/BowlingFramesTests.java
@@ -1,6 +1,5 @@
-package bowling;
+package bowling.frame;
 
-import bowling.frame.BowlingFrames;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -10,7 +9,7 @@ import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-@DisplayName("Bowling Frames Tests")
+@DisplayName(" Tests")
 public class BowlingFramesTests {
 
     @DisplayName("생성 테스트")
