+++ src/test/java/bowling/frame/CommonBowlingFrameTests.java
@@ -1,4 +1,4 @@
-package bowling;
+package bowling.frame;
 
 import bowling.frame.CommonBowlingFrame;
 import org.junit.jupiter.api.DisplayName;
