+++ src/test/java/bowling/framestate/common/FirstBowlTests.java
@@ -1,0 +1,4 @@
+package bowling.framestate.common;
+
+public class FirstBowlTests {
+}
