+++ src/test/java/bowling/framestate/common/MissTests.java
@@ -1,0 +1,4 @@
+package bowling.framestate.common;
+
+public class MissTests {
+}
