+++ src/test/java/bowling/framestate/last/SpareLastFrameOverTests.java
@@ -1,0 +1,4 @@
+package bowling.framestate.last;
+
+public class SpareLastFrameOverTests {
+}
