+++ src/test/java/bowling/framestate/last/StrikeLastFrameTests.java
@@ -1,0 +1,4 @@
+package bowling.framestate.last;
+
+public class StrikeLastFrameTests {
+}
